requirements.txt
# MCP Core Dependencies
mcp>=1.0.0
pydantic>=2.0.0

# HTTP and API clients
httpx>=0.25.0
aiohttp>=3.8.0
requests>=2.31.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Caching and performance
redis>=5.0.0
# aioredis>=2.0.0  # Optional - has Python 3.13 compatibility issues
cachetools>=5.3.0

# RSS and feed parsing
feedparser>=6.0.0
beautifulsoup4>=4.12.0

# Data validation and parsing
xmltodict>=0.13.0
python-dateutil>=2.8.0
pytz>=2023.3

# Optional dependencies for specific APIs
python-whois>=0.8.0
yfinance>=0.2.0

# Logging and monitoring
structlog>=23.1.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0 